requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
pydantic
python-dotenv
